Moving categories and lookups to their own modules
