+++ lib/glimesh/stream_layout/followers_homepage.ex
@@ -2,53 +2,53 @@ defmodule Glimesh.StreamLayout.FollowersHomepage do
   @moduledoc """
   FollowersHomepage is responsible for building an a list of streams that are live.
   """
-  import GlimeshWeb.Gettext
+  # import GlimeshWeb.Gettext
 
-  alias Glimesh.StreamLayout.PageSection
-  alias Glimesh.Streams
+  # alias Glimesh.StreamLayout.PageSection
+  # alias Glimesh.Streams
 
-  def generate_following_page(user) do
-    new_page()
-    |> set_title()
-    |> set_section(user)
-    |> set_layout()
-  end
+  # def generate_following_page(user) do
+  #   new_page()
+  #   |> set_title()
+  #   |> set_section(user)
+  #   |> set_layout()
+  # end
 
-  def new_page do
-    %{}
-  end
+  # def new_page do
+  #   %{}
+  # end
 
-  def set_title(page) do
-    Map.put(page, :title, gettext("Following"))
-  end
+  # def set_title(page) do
+  #   Map.put(page, :title, gettext("Following"))
+  # end
 
-  def set_section(page, user) do
-    channels = Streams.list_live_followed_channels(user)
+  # def set_section(page, user) do
+  #   channels = Streams.list_live_followed_channels(user)
 
-    sections = [
-      %PageSection{
-        # Title of the section
-        title: gettext("Your Followed Streams"),
-        # How the category should show up, eg: half, full
-        layout: "full",
-        bs_parent_class: "col-md-12",
-        bs_child_class: "col-md-3",
+  #   sections = [
+  #     %PageSection{
+  #       # Title of the section
+  #       title: gettext("Your Followed Streams"),
+  #       # How the category should show up, eg: half, full
+  #       layout: "full",
+  #       bs_parent_class: "col-md-12",
+  #       bs_child_class: "col-md-3",
 
-        # Streams that should be shown in order
-        channels: channels
-      }
-    ]
+  #       # Streams that should be shown in order
+  #       channels: channels
+  #     }
+  #   ]
 
-    Map.put(page, :sections, sections)
-  end
+  #   Map.put(page, :sections, sections)
+  # end
 
-  def set_layout(page) do
-    layout =
-      case page.sections do
-        [] -> "none"
-        _ -> "category-top"
-      end
+  # def set_layout(page) do
+  #   layout =
+  #     case page.sections do
+  #       [] -> "none"
+  #       _ -> "category-top"
+  #     end
 
-    Map.put(page, :layout, layout)
-  end
+  #   Map.put(page, :layout, layout)
+  # end
 end
