+++ lib/glimesh/streams.ex
@@ -6,14 +6,14 @@ defmodule Glimesh.Streams do
   import Ecto.Query, warn: false
   alias Glimesh.Accounts.User
   alias Glimesh.Repo
-  alias Glimesh.Streams.Category
   alias Glimesh.Streams.Channel
   alias Glimesh.Streams.Followers
 
   alias Glimesh.Streams.StreamMetadata
   alias Glimesh.Streams.Tag
 
-  alias Glimesh.Streams.Category
+  alias Glimesh.ChannelLookups
+  alias Glimesh.ChannelCategories
 
   defdelegate authorize(action, user, params), to: Glimesh.Streams.Policy
 
@@ -48,7 +48,10 @@ defmodule Glimesh.Streams do
 
   # User API Calls
 
-  def create_channel(%User{} = user, attrs \\ %{category_id: Enum.at(list_categories(), 0).id}) do
+  def create_channel(
+        %User{} = user,
+        attrs \\ %{category_id: Enum.at(ChannelCategories.list_categories(), 0).id}
+      ) do
     with :ok <- Bodyguard.permit(__MODULE__, :create_channel, user) do
       %Channel{
         user: user
@@ -96,7 +99,7 @@ defmodule Glimesh.Streams do
   defp upsert_channel_category_tags(category_id, tag_list) do
     Enum.map(tag_list, fn tag_name ->
       {:ok, tag} =
-        upsert_tag(%Tag{
+        ChannelCategories.upsert_tag(%Tag{
           name: tag_name,
           category_id: category_id
         })
@@ -122,108 +125,6 @@ defmodule Glimesh.Streams do
       |> Channel.changeset(attrs)
       |> Repo.update()
     end
-  end
-
-  ## Categories
-
-  @doc """
-  Returns the list of categories.
-
-  ## Examples
-
-      iex> list_categories()
-      [%Category{}, ...]
-
-  """
-  def list_categories do
-    Repo.all(Category)
-  end
-
-  def list_categories_for_select do
-    Repo.all(from c in Category, order_by: [asc: :name])
-    |> Enum.map(&{&1.name, &1.id})
-  end
-
-  @doc """
-  Gets a single category.
-
-  ## Examples
-
-      iex> get_category(123)
-      %Category{}
-
-      iex> get_category(456)
-      nil
-
-  """
-  def get_category(slug),
-    do: Repo.one(from c in Category, where: c.slug == ^slug)
-
-  def get_category_by_id!(id), do: Repo.get_by!(Category, id: id)
-
-  @doc """
-  Creates a category.
-
-  ## Examples
-
-      iex> create_category(%{field: value})
-      {:ok, %Category{}}
-
-      iex> create_category(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_category(attrs \\ %{}) do
-    %Category{}
-    |> Category.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a category.
-
-  ## Examples
-
-      iex> update_category(category, %{field: new_value})
-      {:ok, %Category{}}
-
-      iex> update_category(category, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_category(%Category{} = category, attrs) do
-    category
-    |> Category.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a category.
-
-  ## Examples
-
-      iex> delete_category(category)
-      {:ok, %Category{}}
-
-      iex> delete_category(category)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_category(%Category{} = category) do
-    Repo.delete(category)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking category changes.
-
-  ## Examples
-
-      iex> change_category(category)
-      %Ecto.Changeset{data: %Category{}}
-
-  """
-  def change_category(%Category{} = category, attrs \\ %{}) do
-    Category.changeset(category, attrs)
   end
 
   # Streams
@@ -263,7 +164,7 @@ defmodule Glimesh.Streams do
         |> Repo.update()
 
       # 4. Send Notifications
-      users = list_live_subscribed_followers(channel)
+      users = ChannelLookups.list_live_subscribed_followers(channel)
 
       Glimesh.Streams.ChannelNotifier.deliver_live_channel_notifications(
         users,
@@ -304,7 +205,7 @@ defmodule Glimesh.Streams do
         ended_at: DateTime.utc_now() |> DateTime.to_naive()
       })
 
-    get_channel!(stream.channel_id)
+    ChannelLookups.get_channel!(stream.channel_id)
     |> Channel.stop_changeset(%{})
     |> Repo.update()
 
@@ -369,7 +270,7 @@ defmodule Glimesh.Streams do
       |> Followers.changeset(attrs)
       |> Repo.insert()
 
-    channel = get_channel_for_user(streamer)
+    channel = ChannelLookups.get_channel_for_user(streamer)
 
     if !is_nil(channel) and Glimesh.Chat.can_create_chat_message?(channel, user) do
       Glimesh.Chat.create_chat_message(user, channel, %{
@@ -413,67 +314,6 @@ defmodule Glimesh.Streams do
     Channel.changeset(channel, attrs)
   end
 
-  def list_channels do
-    Repo.all(
-      from c in Channel,
-        join: cat in Category,
-        on: cat.id == c.category_id
-    )
-    |> Repo.preload([:category, :user])
-  end
-
-  def list_live_channels do
-    Repo.all(
-      from c in Channel,
-        where: c.status == "live"
-    )
-    |> Repo.preload([:category, :user, :stream])
-  end
-
-  def list_in_category(category) do
-    Repo.all(
-      from c in Channel,
-        join: cat in Category,
-        on: cat.id == c.category_id,
-        where: c.status == "live",
-        where: cat.id == ^category.id
-    )
-    |> Repo.preload([:category, :user, :stream, :tags])
-  end
-
-  def filter_live_channels(params) do
-    Repo.all(actually_filter_live_channels(params))
-    |> Repo.preload([:category, :user, :stream, :tags])
-  end
-
-  def actually_filter_live_channels(%{"category" => _category_slug, "tag" => tag_slug}) do
-    from c in Channel,
-      join: t in Tag,
-      join: ct in "channel_tags",
-      on: ct.tag_id == t.id and ct.channel_id == c.id,
-      where: c.status == "live" and t.slug == ^tag_slug
-  end
-
-  def actually_filter_live_channels(%{"category" => category_slug}) do
-    from c in Channel,
-      join: cat in Category,
-      on: cat.id == c.category_id,
-      where: c.status == "live",
-      where: cat.slug == ^category_slug
-  end
-
-  def list_live_subscribed_followers(%Channel{} = channel) do
-    Repo.all(
-      from u in User,
-        left_join: f in Followers,
-        on: u.id == f.user_id,
-        where:
-          f.streamer_id == ^channel.user_id and
-            f.has_live_notifications == true and
-            u.allow_live_subscription_emails == true
-    )
-  end
-
   def list_all_follows do
     Repo.all(from(f in Followers))
   end
@@ -486,219 +326,9 @@ defmodule Glimesh.Streams do
     Repo.all(from f in Followers, where: f.user_id == ^user.id)
   end
 
-  def list_live_followed_channels(user) do
-    Repo.all(
-      from c in Channel,
-        join: f in Followers,
-        on: c.user_id == f.streamer_id,
-        where: c.status == "live",
-        where: f.user_id == ^user.id
-    )
-    |> Repo.preload([:category, :user, :stream, :tags])
-  end
-
-  def list_all_followed_channels(user) do
-    Repo.all(
-      from c in Channel,
-        join: f in Followers,
-        on: c.user_id == f.streamer_id,
-        where: f.user_id == ^user.id
-    )
-    |> Repo.preload([:category, :user, :stream])
-  end
-
-  def list_followed_live_notification_channels(user) do
-    Repo.all(
-      from c in Channel,
-        join: f in Followers,
-        on: c.user_id == f.streamer_id,
-        where: f.user_id == ^user.id and f.has_live_notifications == true
-    )
-    |> Repo.preload([:category, :user, :streamer])
-  end
-
-  def get_channel(id) do
-    Repo.get_by(Channel, id: id) |> Repo.preload([:category, :user, :tags])
-  end
-
-  def get_channel!(id) do
-    Repo.get_by!(Channel, id: id) |> Repo.preload([:category, :user, :tags])
-  end
-
-  def get_channel_for_username!(username, ignore_banned \\ false) do
-    query =
-      Channel
-      |> join(:inner, [u], assoc(u, :user), as: :user)
-      |> where([user: u], u.username == ^username)
-      |> where([c], c.inaccessible == false)
-
-    query =
-      if ignore_banned do
-        query
-      else
-        where(query, [user: u], u.is_banned == false)
-      end
-
-    Repo.one(query)
-    |> Repo.preload([:category, :user, :tags])
-  end
-
-  def get_channel_for_stream_key!(stream_key) do
-    Repo.one(
-      from c in Channel,
-        where: c.stream_key == ^stream_key and c.inaccessible == false
-    )
-    |> Repo.preload([:category, :user])
-  end
-
-  def get_channel_for_user(user) do
-    Repo.one(
-      from c in Channel,
-        join: u in User,
-        on: c.user_id == u.id,
-        where: u.id == ^user.id,
-        where: c.inaccessible == false
-    )
-    |> Repo.preload([:category, :user])
-  end
-
   def is_live?(%Channel{} = channel) do
     channel.status == "live"
   end
 
   # Private Calls
-
-  @doc """
-  Returns the list of tags.
-
-  ## Examples
-
-      iex> list_tags()
-      [%Tag{}, ...]
-
-  """
-  def list_tags do
-    Repo.all(from t in Tag, order_by: [desc: :count_usage]) |> Repo.preload(:category)
-  end
-
-  def list_tags(category_id) do
-    Repo.all(from t in Tag, where: t.category_id == ^category_id, order_by: [desc: :count_usage])
-  end
-
-  def list_live_tags(category_id) do
-    Repo.all(
-      from t in Tag,
-        join: c in Channel,
-        join: ct in "channel_tags",
-        on: ct.tag_id == t.id and ct.channel_id == c.id,
-        where: c.status == "live" and t.category_id == ^category_id,
-        order_by: [desc: :count_usage]
-    )
-  end
-
-  def list_tags_for_tagify(category_id) do
-    Repo.all(
-      from t in Tag,
-        where: t.category_id == ^category_id,
-        order_by: [desc: :count_usage]
-    )
-    |> Enum.map(fn tag ->
-      %{
-        value: tag.name,
-        label: "#{tag.name} (#{tag.count_usage} Uses)",
-        # placeholder for global tags
-        class: ""
-      }
-    end)
-  end
-
-  @doc """
-  Gets a single tag.
-
-  Raises `Ecto.NoResultsError` if the Tag does not exist.
-
-  ## Examples
-
-      iex> get_tag!(123)
-      %Tag{}
-
-      iex> get_tag!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_tag!(id), do: Repo.get!(Tag, id)
-
-  @doc """
-  Creates a tag.
-
-  ## Examples
-
-      iex> create_tag(%{field: value})
-      {:ok, %Tag{}}
-
-      iex> create_tag(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_tag(attrs \\ %{}) do
-    upsert_tag(%Tag{}, attrs)
-  end
-
-  @doc """
-  Updates a tag.
-
-  ## Examples
-
-      iex> update_tag(tag, %{field: new_value})
-      {:ok, %Tag{}}
-
-      iex> update_tag(tag, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_tag(%Tag{} = tag, attrs) do
-    # upsert_tag(tag, attrs)
-    tag
-    |> Tag.changeset(attrs)
-    |> Repo.update()
-  end
-
-  def upsert_tag(%Tag{} = tag, attrs \\ %{}) do
-    tag
-    |> Tag.changeset(attrs)
-    |> Repo.insert(
-      returning: [:count_usage],
-      on_conflict: [inc: [count_usage: 1]],
-      conflict_target: [:identifier]
-    )
-  end
-
-  @doc """
-  Deletes a tag.
-
-  ## Examples
-
-      iex> delete_tag(tag)
-      {:ok, %Tag{}}
-
-      iex> delete_tag(tag)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_tag(%Tag{} = tag) do
-    Repo.delete(tag)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking tag changes.
-
-  ## Examples
-
-      iex> change_tag(tag)
-      %Ecto.Changeset{data: %Tag{}}
-
-  """
-  def change_tag(%Tag{} = tag, attrs \\ %{}) do
-    Tag.changeset(tag, attrs)
-  end
 end
