+++ lib/glimesh/workers/stream_pruner.ex
@@ -6,6 +6,7 @@ defmodule Glimesh.Workers.StreamPruner do
 
   require Logger
   alias Glimesh.Streams
+  alias Glimesh.ChannelLookups
 
   # 5 Minutes in milliseconds
   @interval 300_000
@@ -29,7 +30,7 @@ defmodule Glimesh.Workers.StreamPruner do
   end
 
   defp prune_old_streams do
-    channels = Streams.list_live_channels()
+    channels = ChannelLookups.list_live_channels()
     Logger.info("Checking for stale streams to prune")
 
     current_time = NaiveDateTime.local_now()
