+++ lib/glimesh_web/live/admin/category_live/index.ex
@@ -1,7 +1,7 @@
 defmodule GlimeshWeb.Admin.CategoryLive.Index do
   use GlimeshWeb, :live_view
 
-  alias Glimesh.Streams
+  alias Glimesh.ChannelCategories
   alias Glimesh.Streams.Category
 
   @impl true
@@ -17,7 +17,7 @@ defmodule GlimeshWeb.Admin.CategoryLive.Index do
   defp apply_action(socket, :edit, %{"id" => id}) do
     socket
     |> put_page_title(gettext("Edit Category"))
-    |> assign(:category, Streams.get_category_by_id!(id))
+    |> assign(:category, ChannelCategories.get_category_by_id!(id))
   end
 
   defp apply_action(socket, :new, _params) do
@@ -34,13 +34,13 @@ defmodule GlimeshWeb.Admin.CategoryLive.Index do
 
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
-    category = Streams.get_category_by_id!(id)
-    {:ok, _} = Streams.delete_category(category)
+    category = ChannelCategories.get_category_by_id!(id)
+    {:ok, _} = ChannelCategories.delete_category(category)
 
     {:noreply, assign(socket, :categories, list_categories())}
   end
 
   defp list_categories do
-    Streams.list_categories()
+    ChannelCategories.list_categories()
   end
 end
