+++ lib/glimesh_web/live/admin/category_live/show.ex
@@ -1,7 +1,7 @@
 defmodule GlimeshWeb.Admin.CategoryLive.Show do
   use GlimeshWeb, :live_view
 
-  alias Glimesh.Streams
+  alias Glimesh.ChannelCategories
 
   @impl true
   def mount(_params, _session, socket) do
@@ -13,7 +13,7 @@ defmodule GlimeshWeb.Admin.CategoryLive.Show do
     {:noreply,
      socket
      |> put_page_title(page_title(socket.assigns.live_action))
-     |> assign(:category, Streams.get_category_by_id!(id))}
+     |> assign(:category, ChannelCategories.get_category_by_id!(id))}
   end
 
   defp page_title(:show), do: gettext("Show Category")
