+++ lib/glimesh_web/live/streams_live/list.ex
@@ -3,6 +3,8 @@ defmodule GlimeshWeb.StreamsLive.List do
 
   alias Glimesh.Accounts
   alias Glimesh.Streams
+  alias Glimesh.ChannelLookups
+  alias Glimesh.ChannelCategories
 
   @impl true
   def mount(%{"category" => "following"}, session, socket) do
@@ -10,7 +12,7 @@ defmodule GlimeshWeb.StreamsLive.List do
       %Glimesh.Accounts.User{} = user ->
         if session["locale"], do: Gettext.put_locale(session["locale"])
 
-        channels = Streams.list_live_followed_channels(user)
+        channels = ChannelLookups.list_live_followed_channels(user)
 
         {:ok,
          socket
@@ -28,9 +30,9 @@ defmodule GlimeshWeb.StreamsLive.List do
   def mount(params, session, socket) do
     if session["locale"], do: Gettext.put_locale(session["locale"])
 
-    case Streams.get_category(params["category"]) do
+    case ChannelCategories.get_category(params["category"]) do
       %Streams.Category{} = category ->
-        tags = Streams.list_live_tags(category.id)
+        tags = ChannelCategories.list_live_tags(category.id)
 
         {:ok,
          socket
@@ -47,7 +49,7 @@ defmodule GlimeshWeb.StreamsLive.List do
 
   @impl true
   def handle_params(params, _uri, socket) do
-    channels = Streams.filter_live_channels(Map.take(params, ["category", "tag"]))
+    channels = ChannelLookups.filter_live_channels(Map.take(params, ["category", "tag"]))
 
     {:noreply,
      socket
