+++ lib/glimesh_web/templates/layout/_navbar.html.leex
@@ -63,7 +63,7 @@
                     <% end %>
                     <% end %>
                     <!-- Check for a channel and display an option to go to it -->
-                    <%= if Glimesh.Streams.get_channel_for_user(@current_user) do %>
+                    <%= if Glimesh.ChannelLookups.get_channel_for_user(@current_user) do %>
                     <%= link class: "dropdown-item", to: Routes.user_stream_path(@conn, :index, @current_user.username) do %>
                     <i class="fas fa-tv fa-fw"></i>
                     <%= gettext("Channel") %>
