+++ test/glimesh/channel_lookups_test.exs
@@ -1,0 +1,8 @@
+defmodule Glimesh.ChannelLookupsTest do
+  use Glimesh.DataCase
+  use Bamboo.Test
+
+  #   import Glimesh.AccountsFixtures
+
+  # Coming Soon
+end
