+++ test/glimesh/streams_test.exs
@@ -4,12 +4,9 @@ defmodule Glimesh.StreamsTest do
 
   import Glimesh.AccountsFixtures
   alias Glimesh.Streams
+  alias Glimesh.ChannelLookups
 
   describe "followers" do
-    @valid_attrs %{has_live_notifications: true}
-    @update_attrs %{has_live_notifications: false}
-    @invalid_attrs %{has_live_notifications: nil}
-
     def followers_fixture do
       streamer = streamer_fixture()
       user = user_fixture()
@@ -24,7 +21,7 @@ defmodule Glimesh.StreamsTest do
       user = user_fixture()
       Streams.follow(streamer, user)
 
-      followed = Streams.list_all_followed_channels(user)
+      followed = ChannelLookups.list_all_followed_channels(user)
 
       assert Enum.map(followed, fn x -> x.user.username end) == [streamer.username]
     end
@@ -33,12 +30,12 @@ defmodule Glimesh.StreamsTest do
       streamer = streamer_fixture()
       user = user_fixture()
       Streams.follow(streamer, user)
-      followed = Streams.list_all_followed_channels(user)
+      followed = ChannelLookups.list_all_followed_channels(user)
 
       assert Enum.map(followed, fn x -> x.user.username end) == [streamer.username]
 
       Streams.unfollow(streamer, user)
-      assert Streams.list_all_followed_channels(user) == []
+      assert ChannelLookups.list_all_followed_channels(user) == []
     end
 
     test "is_following?/1 detects active follow" do
@@ -54,71 +51,6 @@ defmodule Glimesh.StreamsTest do
 
       Streams.follow(streamer, user)
       assert {:error, %Ecto.Changeset{}} = Streams.follow(streamer, user)
-    end
-  end
-
-  describe "categories" do
-    alias Glimesh.Streams.Category
-
-    @valid_attrs %{
-      name: "some name"
-    }
-    @update_attrs %{
-      name: "some updated name"
-    }
-    @invalid_attrs %{name: nil}
-
-    def category_fixture(attrs \\ %{}) do
-      {:ok, category} =
-        attrs
-        |> Enum.into(@valid_attrs)
-        |> Streams.create_category()
-
-      category
-    end
-
-    test "list_categories/0 returns all categories" do
-      category = category_fixture()
-      assert Enum.member?(Enum.map(Streams.list_categories(), fn x -> x.name end), category.name)
-    end
-
-    test "get_category_by_id!/1 returns the category with given id" do
-      category = category_fixture()
-      assert Streams.get_category_by_id!(category.id) == category
-    end
-
-    test "create_category/1 with valid data creates a category" do
-      assert {:ok, %Category{} = category} = Streams.create_category(@valid_attrs)
-      assert category.name == "some name"
-      assert category.slug == "some-name"
-    end
-
-    test "create_category/1 with invalid data returns error changeset" do
-      assert {:error, %Ecto.Changeset{}} = Streams.create_category(@invalid_attrs)
-    end
-
-    test "update_category/2 with valid data updates the category" do
-      category = category_fixture()
-      assert {:ok, %Category{} = category} = Streams.update_category(category, @update_attrs)
-      assert category.name == "some updated name"
-      assert category.slug == "some-updated-name"
-    end
-
-    test "update_category/2 with invalid data returns error changeset" do
-      category = category_fixture()
-      assert {:error, %Ecto.Changeset{}} = Streams.update_category(category, @invalid_attrs)
-      assert category == Streams.get_category_by_id!(category.id)
-    end
-
-    test "delete_category/1 deletes the category" do
-      category = category_fixture()
-      assert {:ok, %Category{}} = Streams.delete_category(category)
-      assert_raise Ecto.NoResultsError, fn -> Streams.get_category_by_id!(category.id) end
-    end
-
-    test "change_category/1 returns a category changeset" do
-      category = category_fixture()
-      assert %Ecto.Changeset{} = Streams.change_category(category)
     end
   end
 
@@ -143,7 +75,7 @@ defmodule Glimesh.StreamsTest do
 
     test "start_stream/1 successfully starts a stream", %{channel: channel} do
       {:ok, stream} = Streams.start_stream(channel)
-      new_channel = Streams.get_channel!(channel.id)
+      new_channel = ChannelLookups.get_channel!(channel.id)
 
       assert stream.started_at != nil
       assert stream.ended_at == nil
@@ -172,9 +104,9 @@ defmodule Glimesh.StreamsTest do
 
     test "end_stream/1 successfully stops a stream", %{channel: channel} do
       {:ok, _} = Streams.start_stream(channel)
-      fresh_channel = Streams.get_channel!(channel.id)
+      fresh_channel = ChannelLookups.get_channel!(channel.id)
       {:ok, stream} = Streams.end_stream(fresh_channel)
-      new_channel = Streams.get_channel!(channel.id)
+      new_channel = ChannelLookups.get_channel!(channel.id)
 
       assert stream.started_at != nil
       assert stream.ended_at != nil
@@ -185,7 +117,7 @@ defmodule Glimesh.StreamsTest do
     test "end_stream/1 successfully stops a stream with stream", %{channel: channel} do
       {:ok, stream} = Streams.start_stream(channel)
       {:ok, stream} = Streams.end_stream(stream)
-      new_channel = Streams.get_channel!(channel.id)
+      new_channel = ChannelLookups.get_channel!(channel.id)
 
       assert stream.started_at != nil
       assert stream.ended_at != nil
@@ -214,89 +146,6 @@ defmodule Glimesh.StreamsTest do
       }
 
       assert {:ok, %{}} = Streams.log_stream_metadata(stream, incoming_attrs)
-    end
-  end
-
-  describe "tags" do
-    alias Glimesh.Streams.Tag
-
-    @valid_attrs %{
-      count_usage: 42,
-      name: "some name"
-    }
-    @update_attrs %{
-      count_usage: 43,
-      name: "some updated name"
-    }
-    @invalid_attrs %{category_id: nil, count_usage: nil, icon: nil, name: nil, slug: nil}
-
-    def tag_fixture(attrs \\ %{}) do
-      {:ok, tag} =
-        attrs
-        |> Enum.into(@valid_attrs)
-        |> Streams.create_tag()
-
-      tag
-    end
-
-    test "list_tags/0 returns all tags" do
-      tag = tag_fixture()
-      assert Enum.member?(Enum.map(Streams.list_tags(), fn x -> x.name end), tag.name)
-    end
-
-    test "get_tag!/1 returns the tag with given id" do
-      tag = tag_fixture()
-      assert Streams.get_tag!(tag.id) == tag
-    end
-
-    test "create_tag/1 with valid data creates a tag" do
-      assert {:ok, %Tag{} = tag} = Streams.create_tag(@valid_attrs)
-      assert tag.name == "some name"
-      assert tag.slug == "some-name"
-    end
-
-    test "create_tag/1 with invalid data returns error changeset" do
-      assert {:error, %Ecto.Changeset{}} = Streams.create_tag(@invalid_attrs)
-    end
-
-    test "update_tag/2 with valid data updates the tag" do
-      tag = tag_fixture()
-      assert {:ok, %Tag{} = tag} = Streams.update_tag(tag, @update_attrs)
-      assert tag.name == "some updated name"
-    end
-
-    test "update_tag/2 with invalid data returns error changeset" do
-      tag = tag_fixture()
-      assert {:error, %Ecto.Changeset{}} = Streams.update_tag(tag, @invalid_attrs)
-      assert tag == Streams.get_tag!(tag.id)
-    end
-
-    test "upsert_tag/2 creates a tag or increments if exists" do
-      assert {:ok, %Tag{} = tag} = Streams.upsert_tag(%Tag{name: "Hello World"})
-      assert tag.name == "Hello World"
-      assert tag.slug == "hello-world"
-      assert tag.count_usage == 1
-
-      # Test recreating it
-      assert {:ok, %Tag{} = new_tag} = Streams.upsert_tag(%Tag{name: "Hello World"})
-      assert tag.id == new_tag.id
-      assert new_tag.name == "Hello World"
-      assert new_tag.slug == "hello-world"
-      assert new_tag.count_usage == 2
-
-      assert {:ok, %Tag{} = one_more} = Streams.upsert_tag(%Tag{name: "Hello World"})
-      assert one_more.count_usage == 3
-    end
-
-    test "delete_tag/1 deletes the tag" do
-      tag = tag_fixture()
-      assert {:ok, %Tag{}} = Streams.delete_tag(tag)
-      assert_raise Ecto.NoResultsError, fn -> Streams.get_tag!(tag.id) end
-    end
-
-    test "change_tag/1 returns a tag changeset" do
-      tag = tag_fixture()
-      assert %Ecto.Changeset{} = Streams.change_tag(tag)
     end
   end
 end
