+++ test/glimesh_web/live/chat_live_test.exs
@@ -5,7 +5,7 @@ defmodule GlimeshWeb.ChatLiveTest do
   import Phoenix.LiveViewTest
 
   alias Glimesh.Chat
-  alias Glimesh.Streams
+  alias Glimesh.ChannelLookups
 
   describe "load chat" do
     setup :register_and_log_in_streamer
@@ -28,7 +28,7 @@ defmodule GlimeshWeb.ChatLiveTest do
     test "old chat messages display", %{conn: conn} do
       conn = generate_proper_conn(conn)
       user = conn.assigns.current_user
-      channel = Streams.get_channel_for_user(user)
+      channel = ChannelLookups.get_channel_for_user(user)
       generate_message_for_channel(user, channel, @valid_chat_message)
 
       {:ok, _view, html} = live(conn, Routes.user_stream_path(conn, :index, user.username))
@@ -39,7 +39,7 @@ defmodule GlimeshWeb.ChatLiveTest do
   describe "mod actions" do
     test "short timeout removes chat message", %{conn: conn} do
       streamer = streamer_fixture()
-      channel = Streams.get_channel_for_user(streamer)
+      channel = ChannelLookups.get_channel_for_user(streamer)
 
       {:ok, chat_message} =
         generate_message_for_removal_test(user_fixture(), streamer, channel, @valid_chat_message)
@@ -60,7 +60,7 @@ defmodule GlimeshWeb.ChatLiveTest do
 
     test "long timeout removes chat message", %{conn: conn} do
       streamer = streamer_fixture()
-      channel = Streams.get_channel_for_user(streamer)
+      channel = ChannelLookups.get_channel_for_user(streamer)
 
       {:ok, chat_message} =
         generate_message_for_removal_test(user_fixture(), streamer, channel, @valid_chat_message)
@@ -81,7 +81,7 @@ defmodule GlimeshWeb.ChatLiveTest do
 
     test "ban removes chat message", %{conn: conn} do
       streamer = streamer_fixture()
-      channel = Streams.get_channel_for_user(streamer)
+      channel = ChannelLookups.get_channel_for_user(streamer)
 
       {:ok, chat_message} =
         generate_message_for_removal_test(user_fixture(), streamer, channel, @valid_chat_message)
@@ -104,7 +104,7 @@ defmodule GlimeshWeb.ChatLiveTest do
   describe "chat preferences" do
     test "toggle timestamps button toggles them", %{conn: conn} do
       user = streamer_fixture()
-      channel = Streams.get_channel_for_user(user)
+      channel = ChannelLookups.get_channel_for_user(user)
 
       {:ok, view, _html} =
         live_isolated(conn, GlimeshWeb.ChatLive.Index,
